requirements.txt
torch>=2.0.0
torchvision
numpy
Pillow
matplotlib
scikit-learn
PyYAML
streamlit
torchtext
